refactor: ExpenseCategoryContainer 리팩토링
