+++ src/app/home/setting/expenseCategory/_components/ExpenseCategoryContainer.type.ts
@@ -1,0 +1,7 @@
+import { ExpenseCategory as ExpenseCategoryInterface } from "./Main.type";
+import { ExpenseCategoryAction } from "@/reducers/expenseCategoryReducer";
+
+export interface ExpenseCategoryContainerProps {
+    expenseCategories: ExpenseCategoryInterface[];
+    expenseCategoriesDispatch: React.Dispatch<ExpenseCategoryAction>;
+}
